+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+# Version 1.0.2
+
+* Fix bug with export types for plug and play environments
+* Don't use default export for singleton instance
+
 # Version 1.0.1
 
 * Initial working implementation